# src/docs/combined-swagger.yaml
openapi: 3.0.0
info:
  title: Synthetix Cache API
  version: 1.0.0
  description: API documentation for the Synthetix Cache API
servers:
  - url: http://localhost:3001
  - url: https://synthetix-cache-api-f7db01015ec1.herokuapp.com
paths:
...
  /perp-account-stats/all/{chain}:
    get:
      summary: Get all perp account stats data
      tags:
        - Perp Account Stats
      parameters:
        - in: path
          name: chain
          schema:
            type: string
          required: false
          description: The chain to get the perp account stats data for
      responses:
        200:
          description: All perp account stats data
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1
                    created_at:
                      type: string
                      format: date-time
                      example: '2024-06-27T14:01:04.001Z'
                    updated_at:
                      type: string
                      format: date-time
                      example: '2024-06-27T14:01:04.001Z'
                    ts:
                      type: string
                      format: date-time
                      example: '2024-06-27T13:00:00.000Z'
                    chain:
                      type: string
                      example: base
                    account_id:
                      type: string
                      example: '0x1234567890123456789012345678901234567890'
                    fees:
                      type: number
                      format: double
                      example: 10.5
                    volume:
                      type: number
                      format: double
                      example: 1000.0
                    amount_liquidated:
                      type: number
                      format: double
                      example: 50.0
                    liquidations:
                      type: integer
                      example: 2
                    cumulative_fees:
                      type: number
                      format: double
                      example: 100.5
                    cumulative_volume:
                      type: number
                      format: double
                      example: 10000.0
        404:
          description: Perp account stats data not found
        500:
          description: Server error
  /perp-account-stats/cumulative-unique-traders/{chain}:
    get:
      summary: Get cumulative unique traders data
      tags:
        - Perp Account Stats
      parameters:
        - in: path
          name: chain
          schema:
            type: string
          required: false
          description: The chain to get the cumulative unique traders data for
      responses:
        200:
          description: The cumulative unique traders data
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: object
                    properties:
                      ts:
                        type: string
                        format: date-time
                        example: '2024-06-27T00:00:00.000Z'
                      cumulative_trader_count:
                        type: integer
                        example: 150
        500:
          description: Server error
  /perp-account-stats/cumulative-unique-traders/summary/{chain}:
    get:
      summary: Get the cumulative unique traders summary statistics
      tags:
        - Perp Account Stats
      parameters:
        - in: path
          name: chain
          schema:
            type: string
          required: true
          description: The chain to get the cumulative unique traders summary statistics for
      responses:
        200:
          description: The cumulative unique traders summary statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  current:
                    type: number
                    format: double
                    example: 1000
                  delta_24h:
                    type: number
                    format: double
                    example: 50
                  delta_7d:
                    type: number
                    format: double
                    example: 200
                  delta_28d:
                    type: number
                    format: double
                    example: 500
                  delta_ytd:
                    type: number
                    format: double
                    example: 800
                  ath:
                    type: number
                    format: double
                    example: 1100
                  atl:
                    type: number
                    format: double
                    example: 100
                  ath_percentage:
                    type: number
                    format: double
                    example: 90.91
                  atl_percentage:
                    type: number
                    format: double
                    example: 1000.00
                  standard_deviation:
                    type: number
                    format: double
                    example: 150.25
        400:
          description: Bad request (invalid or missing chain parameter)
        500:
          description: Server error
  /perp-account-stats/daily-new-unique-traders/{chain}:
    get:
      summary: Get daily new unique traders data
      tags:
        - Perp Account Stats
      parameters:
        - in: path
          name: chain
          schema:
            type: string
          required: true
          description: The chain to get the daily new unique traders data for
      responses:
        200:
          description: Daily new unique traders data
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: object
                    properties:
                      ts:
                        type: string
                        format: date-time
                        example: "2023-05-01T00:00:00Z"
                      daily_new_unique_traders:
                        type: integer
                        example: 42
        400:
          description: Bad request - Invalid chain parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        404:
          description: Daily new unique traders data not found
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /perp-account-stats/daily-new-unique-traders/summary/{chain}:
    get:
      summary: Get summary statistics for daily new unique traders
      tags:
        - Perp Account Stats
      parameters:
        - in: path
          name: chain
          schema:
            type: string
          required: true
          description: The chain to get the summary statistics for
      responses:
        200:
          description: Summary statistics for daily new unique traders
          content:
            application/json:
              schema:
                type: object
                properties:
                  current:
                    type: integer
                    example: 42
                  delta_24h:
                    type: number
                    format: float
                    example: 5.5
                  delta_7d:
                    type: number
                    format: float
                    example: -2.3
                  delta_28d:
                    type: number
                    format: float
                    example: 10.1
                  delta_ytd:
                    type: number
                    format: float
                    example: 15.7
                  ath:
                    type: integer
                    example: 100
                  atl:
                    type: integer
                    example: 5
                  ath_percentage:
                    type: number
                    format: float
                    example: 42.0
                  atl_percentage:
                    type: number
                    format: float
                    example: 840.0
                  standard_deviation:
                    type: number
                    format: float
                    example: 12.3
        400:
          description: Bad request - Invalid or missing chain parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        404:
          description: No data found for the specified chain
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...